+++ shardingsphere-error/shardingsphere-common-error/src/main/java/org/apache/shardingsphere/error/exception/syntax/database/UnknownDatabaseException.java
@@ -15,20 +15,20 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.error.dialect.transaction;
+package org.apache.shardingsphere.error.exception.syntax.database;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
-import org.apache.shardingsphere.error.dialect.SQLDialectException;
+import org.apache.shardingsphere.error.exception.SQLDialectException;
 
 /**
- * Table modify in transaction exception.
+ * Unknown database exception.
  */
 @RequiredArgsConstructor
 @Getter
-public final class TableModifyInTransactionException extends SQLDialectException {
+public final class UnknownDatabaseException extends SQLDialectException {
     
-    private static final long serialVersionUID = 5676889868293244575L;
+    private static final long serialVersionUID = 6442120909016308638L;
     
-    private final String tableName;
+    private final String databaseName;
 }
