+++ shardingsphere-error/shardingsphere-common-error/src/main/java/org/apache/shardingsphere/error/exception/transaction/TableModifyInTransactionException.java
@@ -15,20 +15,20 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.error.dialect.syntax.database;
+package org.apache.shardingsphere.error.exception.transaction;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
-import org.apache.shardingsphere.error.dialect.SQLDialectException;
+import org.apache.shardingsphere.error.exception.SQLDialectException;
 
 /**
- * Unknown database exception.
+ * Table modify in transaction exception.
  */
 @RequiredArgsConstructor
 @Getter
-public final class UnknownDatabaseException extends SQLDialectException {
+public final class TableModifyInTransactionException extends SQLDialectException {
     
-    private static final long serialVersionUID = 6442120909016308638L;
+    private static final long serialVersionUID = 5676889868293244575L;
     
-    private final String databaseName;
+    private final String tableName;
 }
