+++ shardingsphere-error/shardingsphere-mysql-error/src/main/java/org/apache/shardingsphere/error/mysql/mapper/MySQLDialectExceptionMapper.java
@@ -17,16 +17,16 @@
 
 package org.apache.shardingsphere.error.mysql.mapper;
 
-import org.apache.shardingsphere.error.dialect.SQLDialectException;
-import org.apache.shardingsphere.error.dialect.connection.TooManyConnectionsException;
-import org.apache.shardingsphere.error.dialect.data.InsertColumnsAndValuesMismatchedException;
-import org.apache.shardingsphere.error.dialect.syntax.database.DatabaseCreateExistsException;
-import org.apache.shardingsphere.error.dialect.syntax.database.DatabaseDropNotExistsException;
-import org.apache.shardingsphere.error.dialect.syntax.database.NoDatabaseSelectedException;
-import org.apache.shardingsphere.error.dialect.syntax.database.UnknownDatabaseException;
-import org.apache.shardingsphere.error.dialect.syntax.table.NoSuchTableException;
-import org.apache.shardingsphere.error.dialect.syntax.table.TableExistsException;
-import org.apache.shardingsphere.error.dialect.transaction.TableModifyInTransactionException;
+import org.apache.shardingsphere.error.exception.SQLDialectException;
+import org.apache.shardingsphere.error.exception.connection.TooManyConnectionsException;
+import org.apache.shardingsphere.error.exception.data.InsertColumnsAndValuesMismatchedException;
+import org.apache.shardingsphere.error.exception.syntax.database.DatabaseCreateExistsException;
+import org.apache.shardingsphere.error.exception.syntax.database.DatabaseDropNotExistsException;
+import org.apache.shardingsphere.error.exception.syntax.database.NoDatabaseSelectedException;
+import org.apache.shardingsphere.error.exception.syntax.database.UnknownDatabaseException;
+import org.apache.shardingsphere.error.exception.syntax.table.NoSuchTableException;
+import org.apache.shardingsphere.error.exception.syntax.table.TableExistsException;
+import org.apache.shardingsphere.error.exception.transaction.TableModifyInTransactionException;
 import org.apache.shardingsphere.error.mapper.SQLDialectExceptionMapper;
 import org.apache.shardingsphere.error.mysql.code.MySQLVendorError;
 import org.apache.shardingsphere.infra.util.exception.sql.UnknownSQLException;
