+++ shardingsphere-proxy/shardingsphere-proxy-backend/src/main/java/org/apache/shardingsphere/proxy/backend/handler/distsql/ral/updatable/RefreshTableMetadataHandler.java
@@ -23,8 +23,8 @@ import org.apache.shardingsphere.infra.database.type.DatabaseTypeEngine;
 import org.apache.shardingsphere.infra.distsql.exception.DistSQLException;
 import org.apache.shardingsphere.mode.manager.ContextManager;
 import org.apache.shardingsphere.proxy.backend.context.ProxyContext;
-import org.apache.shardingsphere.error.dialect.syntax.database.NoDatabaseSelectedException;
-import org.apache.shardingsphere.error.dialect.syntax.database.UnknownDatabaseException;
+import org.apache.shardingsphere.error.exception.syntax.database.NoDatabaseSelectedException;
+import org.apache.shardingsphere.error.exception.syntax.database.UnknownDatabaseException;
 import org.apache.shardingsphere.proxy.backend.handler.distsql.ral.UpdatableRALBackendHandler;
 
 /**
